需要文字处理部分；暂时改成半精度

+++ model/clip/model.py
@@ -381,21 +381,21 @@ def convert_weights(model: nn.Module):
 
     def _convert_weights_to_fp16(l):  # noqa: E741
         if isinstance(l, (nn.Conv1d, nn.Conv2d, nn.Linear)):
-            l.weight.data = l.weight.data.float()
+            l.weight.data = l.weight.data.half()
             if l.bias is not None:
-                l.bias.data = l.bias.data.float()
+                l.bias.data = l.bias.data.half()
 
         if isinstance(l, nn.MultiheadAttention):
             for attr in [*[f"{s}_proj_weight" for s in ["in", "q", "k", "v"]], "in_proj_bias", "bias_k", "bias_v"]:
                 tensor = getattr(l, attr)
                 if tensor is not None:
-                    tensor.data = tensor.data.float()
+                    tensor.data = tensor.data.half()
 
         for name in ["text_projection", "proj"]:
             if hasattr(l, name):
                 attr = getattr(l, name)
                 if attr is not None:
-                    attr.data = attr.data.float()
+                    attr.data = attr.data.half()
 
     model.apply(_convert_weights_to_fp16)
